Rename ros_timer_action to ros_timer
