+++ launch_ros/launch_ros/actions/__init__.py
@@ -19,7 +19,7 @@ from .lifecycle_node import LifecycleNode
 from .load_composable_nodes import LoadComposableNodes
 from .node import Node
 from .push_ros_namespace import PushRosNamespace
-from .ros_timer_action import RosTimerAction
+from .ros_timer import RosTimer
 from .set_parameter import SetParameter
 from .set_remap import SetRemap
 
@@ -30,7 +30,7 @@ __all__ = [
     'LoadComposableNodes',
     'Node',
     'PushRosNamespace',
-    'RosTimerAction',
+    'RosTimer',
     'SetParameter',
     'SetRemap',
 ]
